+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Fixed
 
 - Ensure CLI console output for config testing is always visible by using standard library log instead of slog ([#113](https://github.com/nite-coder/bifrost/pull/113) by [0x5487](https://github.com/0x5487))
+- Fix missing access log and metrics tracers for HTTP/2 requests in standard library bridge ([#116](https://github.com/nite-coder/bifrost/pull/116) by [0x5487](https://github.com/0x5487))
 
 ## Security